Commands:
$ git fetch origin
$ git checkout -B ui-stable -t origin/ui-stable
$ git merge --strategy=ours origin/ui-canary
$ git diff --binary origin/ui-canary | git apply --reverse --index
$ git commit --amend

End state:
$ git diff ui-stable origin/ui-canary
[no output]
$ git rev-list --count ui-stable..origin/ui-canary
0

+++ ui/src/common/search_data.ts
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 export interface SearchSummary {
-  tsStarts: Float64Array;
-  tsEnds: Float64Array;
+  tsStarts: BigInt64Array;
+  tsEnds: BigInt64Array;
   count: Uint8Array;
 }
 
 export interface CurrentSearchResults {
   sliceIds: Float64Array;
-  tsStarts: Float64Array;
+  tsStarts: BigInt64Array;
   utids: Float64Array;
   trackIds: string[];
   sources: string[];
